涅槃科技 [+]增加服务器 信息/主图 缓存 [+]增加租凭服 信息/主图 缓存 [+]增加皮肤页 信息/主图 缓存 [+]优化了下载逻辑 [优化]......

+++ Nirvana.Public/Entities/IntPtrReference.cs
@@ -1,5 +1,5 @@
 ﻿namespace Nirvana.Public.Entities;
 
-public class IntPtrReference(int initialValue = 100) {
-    public int Value { get; set; } = initialValue;
+public class IntPtrReference(double initialValue = 100) {
+    public double Value { get; set; } = initialValue;
 }
